feat(struct): harden runtime document codec
