+++ src/struct/codec.ts
@@ -1,0 +1,8 @@
+/** Public STRUCT codec entrypoint. */
+export {
+  StructCodecError,
+  decodeStructDocument,
+  encodeStructDocument,
+  migrateStructDocument,
+  type StructDocumentJson,
+} from './codec/index'
