+++ src/struct/codec/index.ts
@@ -1,0 +1,7 @@
+export {
+  StructCodecError,
+  decodeStructDocument,
+  encodeStructDocument,
+  migrateStructDocument,
+  type StructDocumentJson,
+} from './structure'
